refactor internal link previewing into a protocol extension
